Update ImageJ server
The aim is to improve code readability and reduce memory use when opening large images.
